+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 ### General
 - Mol* 3.0.2
 
+### Bug fixes
+- RO-2605: check for single space-group & NCS operator
+
 ## [2.0.9] - 2022-01-24
 ### General
 - Update to Mol* 3.0.0